updated simulation structure
