+++ backend/modules/pdf_processing/repository.py
@@ -124,7 +124,7 @@ class PDFProcessingRepository:
             
             # Save personas - check for existing ones first
             existing_personas = self.db.query(SimulationPersona).filter(
-                SimulationPersona.scenario_id == simulation.id,
+                SimulationPersona.simulation_id == simulation.id,
                 SimulationPersona.deleted_at.is_(None)
             ).all()
             existing_persona_names = {p.name for p in existing_personas}
@@ -141,7 +141,7 @@ class PDFProcessingRepository:
                     traits = figure.get("personality_traits", {}) or figure.get("traits", {})
                     
                     persona = SimulationPersona(
-                        scenario_id=simulation.id,
+                        simulation_id=simulation.id,
                         name=persona_name,
                         role=figure.get("role", ""),
                         background=figure.get("background", ""),
@@ -232,7 +232,7 @@ class PDFProcessingRepository:
             
             # Save personas
             existing_personas = self.db.query(SimulationPersona).filter(
-                SimulationPersona.scenario_id == simulation.id,
+                SimulationPersona.simulation_id == simulation.id,
                 SimulationPersona.deleted_at.is_(None)
             ).all()
             existing_persona_names = {p.name for p in existing_personas}
@@ -247,7 +247,7 @@ class PDFProcessingRepository:
                     traits = figure.get("personality_traits", {}) or figure.get("traits", {})
                     
                     persona = SimulationPersona(
-                        scenario_id=simulation.id,
+                        simulation_id=simulation.id,
                         name=persona_name,
                         role=figure.get("role", ""),
                         background=figure.get("background", ""),
@@ -265,7 +265,7 @@ class PDFProcessingRepository:
             
             # Build persona mapping: name -> id
             all_personas = self.db.query(SimulationPersona).filter(
-                SimulationPersona.scenario_id == simulation.id,
+                SimulationPersona.simulation_id == simulation.id,
                 SimulationPersona.deleted_at.is_(None)
             ).all()
             persona_mapping = {p.name: p.id for p in all_personas}
@@ -292,7 +292,7 @@ class PDFProcessingRepository:
             
             # Save scenes
             existing_scenes = self.db.query(SimulationScene).filter(
-                SimulationScene.scenario_id == simulation.id
+                SimulationScene.simulation_id == simulation.id
             ).all()
             existing_scene_titles = {s.title for s in existing_scenes}
             
@@ -304,7 +304,7 @@ class PDFProcessingRepository:
                         continue
                     
                     scene = SimulationScene(
-                        scenario_id=simulation.id,
+                        simulation_id=simulation.id,
                         title=scene_title,
                         description=scene_data.get("description", ""),
                         user_goal=scene_data.get("user_goal", ""),
@@ -359,12 +359,12 @@ class PDFProcessingRepository:
             
             # Check if images exist
             all_scenes = self.db.query(SimulationScene).filter(
-                SimulationScene.scenario_id == simulation.id
+                SimulationScene.simulation_id == simulation.id
             ).all()
             has_scenes_with_images = any(scene.image_url for scene in all_scenes)
             
             all_personas_final = self.db.query(SimulationPersona).filter(
-                SimulationPersona.scenario_id == simulation.id,
+                SimulationPersona.simulation_id == simulation.id,
                 SimulationPersona.deleted_at.is_(None)
             ).all()
             has_personas_with_images = any(persona.image_url for persona in all_personas_final)
