+++ backend/modules/simulation/core/__init__.py
@@ -1,0 +1,13 @@
+"""Core orchestration and state management for simulations."""
+
+from .state import SimulationState
+from .orchestrator import ChatOrchestrator
+from .orchestrator_manager import OrchestratorManager
+from .scene_progression import SceneProgressionHandler
+
+__all__ = [
+    "SimulationState",
+    "ChatOrchestrator",
+    "OrchestratorManager",
+    "SceneProgressionHandler",
+]
