+++ backend/modules/simulation/handlers/__init__.py
@@ -1,0 +1,5 @@
+"""Chat handlers for simulation module."""
+
+from .chat_handler import ChatHandler
+
+__all__ = ["ChatHandler"]
