+++ backend/modules/simulation/services/progress_service.py
@@ -87,7 +87,7 @@ class ProgressService:
         personas_data = [
             SimulationPersonaResponse(
                 id=p.id,
-                simulation_id=p.scenario_id,
+                simulation_id=p.simulation_id,
                 name=p.name,
                 role=p.role,
                 background=getattr(p, 'background', None),
